CVM Guest VSM: refactor the shared/encrypted bitmaps to be partition-wide (#1370)
In preparation for VTL 1 memory support for CVMs, make the
shared/encrypted bitmap tracking available on a partition-level, rather
than in the GuestMemoryMapping (which ends up being per-VTL). Also
includes some refactoring to isolate out the bitmap logic so that it can
be reused for vtl protection bitmaps.

Tested: SNP +/- guest vsm boots